Add Months to a Date

Naively adding months to a date can produce invalid dates, for example when adding one month to 2012-01-30. addMonths() always returns a valid date: adding one month to as.Date("2013-01-31") returns "2013-02-28". Negative values of n subtract months.

addMonths(x, n, ...)

Arguments

x

a Date object, or an object coercible to one by as.Date(x, ...), to which months are added

n

number of months to add. Negative values subtract months.

...

further arguments passed to as.Date, for example to supply origin

Value

a vector of class Date with the same dimension as x, containing the transformed dates

Details

All parameters are recycled if necessary.

Note

Based on code by Roland Rapold and Antonio, adapted to conform to package standards.

References

https://stackoverflow.com/questions/14169620/add-a-month-to-a-date

See also

Examples

# characters will be coerced to Date
addMonths("2013-01-31", 1)
#> [1] "2013-02-28"

# negative n
addMonths(as.Date("2013-03-31"), -1)
#> [1] "2013-02-28"

# Arguments will be recycled
# (with warning if the longer is not a multiple of length of shorter)
addMonths(c("2013-01-31", "2013-03-31", "2013-10-31", "2013-12-31"), c(1,-1))
#> [1] "2013-02-28" "2013-02-28" "2013-11-30" "2013-11-30"

x <- as.POSIXct(c("2015-01-31", "2015-08-31"))
n <- c(1, 3)
addMonths(x, n)
#> [1] "2015-02-28" "2015-11-30"

# mind the origin if x supplied as numeric ...
x <- as.numeric(as.Date(x))
addMonths(x, n, origin=as.Date("1970-01-01"))
#> [1] "2015-02-28" "2015-11-30"