Cohen's and Hedges' Effect Size

Computes the Cohen's d and Hedges' g effect size statistics.

cohenD(
  x,
  y = NULL,
  conf.level = NA,
  sides = c("two.sided", "left", "right"),
  correct = FALSE,
  na.rm = FALSE
)

Arguments

x

a non-empty numeric vector of data values

y

an optional non-empty numeric vector of data values

conf.level

confidence level of the interval. Set this to NA if no confidence interval should be calculated. This is the default.

sides

a character string specifying the side of the confidence interval. Must be one of "two.sided" (default), "left", or "right". Partial matching is supported. "left" is analogous to a hypothesis of "greater" in a t.test.

correct

logical; whether to apply the Hedges correction. Defaults to FALSE.

na.rm

logical. Should missing values be removed? Defaults to FALSE.

Value

if conf.level = NA, a numeric scalar containing the effect size; otherwise a named numeric vector with elements:

est

point estimate of Cohen's \(d\) or Hedges' \(g\).

lci

lower confidence interval bound.

uci

upper confidence interval bound.

The magnitude category and pooled standard deviation are stored in the attributes magnitude and sd_pooled, respectively.

Note

Based on code by William Revelle.

References

Cohen, J. (1988) Statistical power analysis for the behavioral sciences (2nd ed.) Academic Press, New York.

Hedges, L. V. & Olkin, I. (1985) Statistical methods for meta-analysis Academic Press, Orlando, FL

Smithson, M.J. (2003) Confidence Intervals, Quantitative Applications in the Social Sciences Series, No. 140. Thousand Oaks, CA: Sage. pp. 39-41

See also

Examples

x <- Pizza$price[Pizza$driver=="Carter"]
y <- Pizza$price[Pizza$driver=="Miller"]

cohenD(x, y, conf.level=0.95, na.rm=TRUE)
#>          est          lci          uci 
#> -0.212277884 -0.430317413  0.006061792 
#> attr(,"magnitude")
#> [1] "small"
#> attr(,"sd_pooled")
#> [1] 21.54513