Cramer's V

Measures the strength of association between two categorical variables. These can be provided as two data vectors x and y, or as a contingency table (see Association).

cramerV(
  x,
  y = NULL,
  conf.level = NA,
  method = c("ncchisq", "ncchisqadj", "fisher", "fisheradj"),
  correct = FALSE,
  ...
)

Arguments

x

either a contingency table, a two-column object (matrix, data.frame or list), or a vector of observations (together with y)

y

optional second vector. If x is not a vector, y must be NULL.

conf.level

confidence level for confidence intervals. If NA, no confidence interval is computed.

method

character string specifying the confidence interval method: "ncchisq" (default, using the noncentral chi-squared distribution), "ncchisqadj", "fisher" (using fisher z transformation), or "fisheradj" (using the Fisher z transformation and bias correction)

correct

logical; whether to apply the bias correction of Bergsma (2013); defaults to FALSE

...

additional arguments passed to internal helper functions, such as bootstrap routines (boot) or table handling (table) (e.g. useNA, R)

Value

if conf.level = NA, a numeric scalar containing Cramer's V; otherwise a named numeric vector with elements:

est

point estimate of Cramer's V.

lci

lower confidence interval bound.

uci

upper confidence interval bound.

Details

Cramer's V ranges from 0 to 1, with 0 indicating statistical independence.

Note

Based on code by Michael Smithson (confidence intervals), adapted to conform to package standards.

References

Cramer, H. (1946) Mathematical Methods of Statistics. Princeton University Press

Agresti, Alan (1996) Introduction to categorical data analysis. NY: John Wiley and Sons

Bergsma, W. (2013) A bias-correction for Cramer's V and Tschuprow's T Journal of the Korean Statistical Society 42(3) DOI: 10.1016/j.jkss.2012.10.002

See also

Examples

tab <- table(Pizza$driver, Pizza$wine_delivered)
cramerV(tab)
#> [1] 0.1328222

# just x and y
cramerV(Pizza$driver, Pizza$wine_delivered)
#> [1] 0.1328222

# data.frame
bedrock::pairApply(Pizza[,c("driver","operator","area")], cramerV, symmetric = TRUE)
#>             driver   operator       area
#> driver   1.0000000 0.23585686 0.65018461
#> operator 0.2358569 1.00000000 0.08670047
#> area     0.6501846 0.08670047 1.00000000

# useNA is passed to table
bedrock::pairApply(Pizza[,c("driver","operator","area")], cramerV,
          useNA="ifany", symmetric = TRUE)
#>             driver   operator       area
#> driver   1.0000000 0.23585686 0.65018461
#> operator 0.2358569 1.00000000 0.08670047
#> area     0.6501846 0.08670047 1.00000000

d.frm <- Pizza[,c("driver","operator","area")]
bedrock::pairApply(d.frm[complete.cases(d.frm),], cramerV, symmetric = TRUE)
#>             driver  operator      area
#> driver   1.0000000 0.2345141 0.6504665
#> operator 0.2345141 1.0000000 0.0869935
#> area     0.6504665 0.0869935 1.0000000

# Bootstrap confidence intervals for Cramer's V
# http://support.sas.com/documentation/cdl/en/statugfreq/63124/PDF/default/statugfreq.pdf, p. 1821

tab <- as.table(rbind(
  c(26,26,23,18, 9),
  c( 6, 7, 9,14,23)))
d.frm <- bedrock::untable(tab)

n <- 1000
idx <- matrix(sample(nrow(d.frm), size=nrow(d.frm) * n, replace=TRUE), ncol=n, byrow=FALSE)
v <- apply(idx, 2, function(x) cramerV(d.frm[x,1], d.frm[x,2]))
quantile(v, probs=c(0.025,0.975))
#>      2.5%     97.5% 
#> 0.2772397 0.5584781 

# compare this to the analytical ones
cramerV(tab, conf.level=0.95)
#>       est       lci       uci 
#> 0.4064888 0.2211575 0.5410535