Describe Relationship: Categorical x by Categorical y

Computes descriptive statistics for the relationship between two categorical variables x and y.

Usage

.descQQ(x, y, ...)

Arguments

x

a categorical variable

y

a categorical variable

...

further arguments, currently unused

Value

an object of class c("Desc.qq", "Desc")

Details

This function is a wrapper around desc.table applied to the contingency table table(x, y).

It summarizes the joint distribution of two categorical variables and provides association measures and visualizations.

Computed statistics

Contingency table

Row and column percentages

Association measures (e.g., Cramer's V, Phi)

Optional statistical tests depending on configuration

Implementation note Internally, desc.qq(x, y) is equivalent to:

desc(table(x, y))

See also