Confidence Intervals for Pearson Correlation

Find the confidence intervals for a specified correlation based on Fisher's z-transformation.

pearsonCor(
  x,
  y = NULL,
  conf.level = NA,
  sides = c("two.sided", "left", "right"),
  method = c("fisher"),
  scoresType = "table",
  na.rm = FALSE
)

Arguments

x

a numeric vector, matrix, or table

y

NULL (default) or a vector with compatible dimensions to x. If y is supplied, table(x, y, ...) is calculated.

conf.level

confidence level of the interval. If set to NA (which is the default) no confidence interval will be calculated.

sides

character string specifying the side of the confidence interval, must be one of "two.sided" (default), "left" or "right". The value names the side on which the finite bound lies. You can specify just the initial letter.

method

confidence interval method; currently only "fisher" is implemented

scoresType

score calculation method for table input

na.rm

logical, default FALSE determining if complete cases should be respected

Value

if conf.level = NA, a numeric scalar. Otherwise a named numeric vector with elements:

est

point estimate of Pearson's correlation coefficient

lci

lower confidence interval bound

uci

upper confidence interval bound

Details

The sampling distribution of Pearson's r is not normal. Fisher developed a transformation now called "Fisher's z-transformation" used for the calculation of normal distributed confidence intervals.

Note

Based on code by William Revelle, adapted to conform to package standards.

See also

Examples

with(swiss, pearsonCor(Fertility, Agriculture))
#> [1] 0.3530792
with(swiss, pearsonCor(Fertility, Agriculture, conf.level=0.95))
#>        est        lci        uci 
#> 0.35307918 0.07334947 0.58130587 

bedrock::pairApply(swiss, pearsonCor)
#>                   Fertility Agriculture Examination   Education   Catholic
#> Fertility         1.0000000  0.35307918  -0.6458827 -0.66378886  0.4636847
#> Agriculture       0.3530792  1.00000000  -0.6865422 -0.63952252  0.4010951
#> Examination      -0.6458827 -0.68654221   1.0000000  0.69841530 -0.5727418
#> Education        -0.6637889 -0.63952252   0.6984153  1.00000000 -0.1538589
#> Catholic          0.4636847  0.40109505  -0.5727418 -0.15385892  1.0000000
#> Infant.Mortality  0.4165560 -0.06085861  -0.1140216 -0.09932185  0.1754959
#>                  Infant.Mortality
#> Fertility              0.41655603
#> Agriculture           -0.06085861
#> Examination           -0.11402160
#> Education             -0.09932185
#> Catholic               0.17549591
#> Infant.Mortality       1.00000000

bedrock::pairApply(swiss, 
           function(x, y) fmCI(pearsonCor(x, y, conf.level=0.95), 
                               digits=3, ldigits=0))
#>                  Fertility                 Agriculture              
#> Fertility        "1.000 [1.000, 1.000]"    "0.353 [0.073, 0.581]"   
#> Agriculture      "0.353 [0.073, 0.581]"    "1.000 [1.000, 1.000]"   
#> Examination      "-0.646 [-0.787, -0.440]" "-0.687 [-0.813, -0.497]"
#> Education        "-0.664 [-0.799, -0.465]" "-0.640 [-0.783, -0.432]"
#> Catholic         "0.464 [0.204, 0.663]"    "0.401 [0.129, 0.617]"   
#> Infant.Mortality "0.417 [0.147, 0.629]"    "-0.061 [-0.342, 0.230]" 
#>                  Examination               Education                
#> Fertility        "-0.646 [-0.787, -0.440]" "-0.664 [-0.799, -0.465]"
#> Agriculture      "-0.687 [-0.813, -0.497]" "-0.640 [-0.783, -0.432]"
#> Examination      "1.000 [1.000, 1.000]"    "0.698 [0.514, 0.821]"   
#> Education        "0.698 [0.514, 0.821]"    "1.000 [1.000, 1.000]"   
#> Catholic         "-0.573 [-0.738, -0.342]" "-0.154 [-0.422, 0.139]" 
#> Infant.Mortality "-0.114 [-0.388, 0.179]"  "-0.099 [-0.376, 0.193]" 
#>                  Catholic                  Infant.Mortality        
#> Fertility        "0.464 [0.204, 0.663]"    "0.417 [0.147, 0.629]"  
#> Agriculture      "0.401 [0.129, 0.617]"    "-0.061 [-0.342, 0.230]"
#> Examination      "-0.573 [-0.738, -0.342]" "-0.114 [-0.388, 0.179]"
#> Education        "-0.154 [-0.422, 0.139]"  "-0.099 [-0.376, 0.193]"
#> Catholic         "1.000 [1.000, 1.000]"    "0.175 [-0.118, 0.440]" 
#> Infant.Mortality "0.175 [-0.118, 0.440]"   "1.000 [1.000, 1.000]"